Time: 0 ms (100.00%) | Memory: 40.4 MB (37.76%) - LeetSync

+++ 476-number-complement/number-complement.java
@@ -1,14 +1,14 @@
 class Solution {
     public int findComplement(int num) {
-        int len = 0;
-        int n = num;
+        long len = 0;
+        long n = num;
         while(n > 0){
             n = n/2 ;
             len++;
         }
-       int setBits =(int) (Math.pow(2,len) -1);
+       long setBits =(long)Math.pow(2,len) -1;
 
-       return num ^ setBits ;
+       return (int)num ^ (int)setBits ;
 
     }
 }
